Use zobristKey for frequency checking
instead of using the fen of the position, use the zobrist key.
This is faster and uses less memory than using the FEN directly, by a sizeable factor.
The disadvantage is that there is a small chance of zobrist key collision, in which case any of the FENs matching the position can be output.
Additionally, if the count of the positions is required in the output, some more memory is required.

An example with data on SSD:
```
$ ./fastpopular --dir download35/ --minCount 2 --maxPlies 60 --stopEarly
Looking for pgn files in download35/
Found 370353 .pgn(.gz) files, creating 128 chunks for processing.
Processed 370353 files
Found 23762247 positions in 496063288 games.
Total time for processing: 55.026s
```
i.e. 10M games / second, games analysed to 60 plies, stopping lines on the first new position, outputting the position if seen at least 2 times.